refactor: create server utils module
